Changed uInt32 to size_t where necessary.
Finally decided to stop fighting C++, and just use size_t everywhere an
array is used.
Even though this is a 64-bit value, it seems that C++/STL expects it
when dealing with arrays, so I guess we will too.

+++ src/debugger/gui/CartDebugWidget.cxx
@@ -37,7 +37,7 @@ CartDebugWidget::CartDebugWidget(GuiObject* boss, const GUI::Font& lfont,
     myDesc(nullptr)   { }
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-int CartDebugWidget::addBaseInformation(int bytes, const string& manufacturer,
+int CartDebugWidget::addBaseInformation(size_t bytes, const string& manufacturer,
         const string& desc, const uInt16 maxlines)
 {
   const int lwidth = _font.getStringWidth("Manufacturer "),
